Labfiles/03c-use-agent-tools-with-mcp/Python/requirements.txt
python-dotenv
azure-identity
azure-ai-agents
uvicorn